analysis: rework the AnalysisRegister object
* Rename it to AnalysisProxy to better reflect the use of its instances
(they are a kind of proxy holding instances of analysis). Also use a
noun instead of a verb in the class name.

* Remove the custom import mechanism and replace it with a simple
inheritance-based plugin system. Python modules can also now hold more
than one analysis class without an issue, and there is no risk of double
import (all analysis modules are imported using one name so they are
cached properly in sys.module dict).

* The list of available analysis class is can be exposed easily, without
confusion with other unrelated attributes of the proxy

* The submodules are imported in __init__.py under the right name, to
avoid double import if they are imported explicitely. That is still
needed to make sure the classes are registered before any instance of
the AnalysisProxy is created.