exekall: improve modules detection
* Restrict the list of testcases to the ones where the final stage of
the step is defined in one of the files given on the command line

* Recursively look for callables in all modules imported from the ones
specified on the command line, in order to have a chance to pick up all
pre-requisites of the given modules. Note that this search is limited to
the packages of the modules specified on the command line, to avoid too
broad results.